Build libspnav so that 3D Mouse support works
In https://github.com/flathub/org.openscad.OpenSCAD/pull/38
changes were made to allow the OpenSCAD Flatpak to access the spacenav
socket, but OpenSCAD wasn't built with support for the library to
consume the data from that socket.

Build libspnav, and tell qmake where the library lives so that 3D
mouse support works.